requirements.txt
blinker==1.4
Bootstrap-Flask==1.0.10
Click==7.0
Faker==1.0.4
Flask==1.0.2
Flask-DebugToolbar==0.10.1
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-dotenv==0.10.1
six==1.12.0
SQLAlchemy==1.3.1
text-unidecode==1.2
Werkzeug==0.14.1
WTForms==2.2.1
